style: rustfmt + split temper-store-sim tests into a module (CI: lint + readability ratchet)
Fixes the two red CI gates on temper#326 (the gates I'd skipped with --no-verify):
- Compile & Lint: cargo fmt across the new key-index code (store_projection_test.rs
  + others had unformatted .await.unwrap() chains).
- Integrity & DST Patterns: the key-index additions pushed temper-store-sim/src/lib.rs
  987 -> 1108 lines, crossing the 1000-line readability ratchet (PROD_FILES_GT1000
  21 -> 22). Extracted the inline #[cfg(test)] mod tests (296 lines) to a sibling
  tests.rs per the repo's >500-line split rule; lib.rs is now 813. Ratchet back to 21.
No production logic changed; sim tests 12/12 green.
